refactor: 미사용 클래스 삭제 및 클래스 패키지 이동(클래스명 수정 포함)
