+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/article/ArticleQueryDslRepository.java
@@ -1,8 +1,8 @@
-package com.mobble.mobbleserver.refactor.article.repository;
+package com.mobble.mobbleserver.infrastructure.persistence.article;
 
-import com.mobble.mobbleserver.refactor.article.entity.Article;
-import com.mobble.mobbleserver.refactor.article.entity.ArticleType;
-import com.mobble.mobbleserver.refactor.article.repository.dto.ArticleLikeInfoDto;
+import com.mobble.mobbleserver.domain.article.Article;
+import com.mobble.mobbleserver.domain.article.ArticleType;
+import com.mobble.mobbleserver.infrastructure.persistence.article.projection.ArticleLikeInfoDto;
 
 import java.util.List;
 import java.util.Map;
