+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/article/projection/ArticleLikeProjection.java
