+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/article/dto/response/ArticleResponseDto.java
@@ -1,8 +1,8 @@
 package com.mobble.mobbleserver.refactor.article.dto.response;
 
 import com.mobble.mobbleserver.infrastructure.web.comment.dto.response.RootCommentResponseDto;
-import com.mobble.mobbleserver.refactor.article.entity.Article;
-import com.mobble.mobbleserver.refactor.article.entity.ArticleType;
+import com.mobble.mobbleserver.domain.article.Article;
+import com.mobble.mobbleserver.domain.article.ArticleType;
 import com.mobble.mobbleserver.refactor.article.repository.dto.ArticleLikeInfoDto;
 
 import java.time.LocalDateTime;
