--- src/main/java/com/mobble/mobbleserver/refactor/article/validator/ArticleValidator.java
@@ -1,25 +1,0 @@
-package com.mobble.mobbleserver.refactor.article.validator;
-
-import com.mobble.mobbleserver.global.exception.common.DomainException;
-import com.mobble.mobbleserver.global.exception.errorCode.article.ArticleErrorCode;
-import com.mobble.mobbleserver.refactor.article.entity.Article;
-import com.mobble.mobbleserver.refactor.article.repository.ArticleRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class ArticleValidator {
-
-    private final ArticleRepository articleRepository;
-
-    public Article findArticleByArticleIdOrThrow(Long articleId) {
-        return articleRepository.findById(articleId)
-                .orElseThrow(() -> new DomainException(ArticleErrorCode.NOT_FOUND));
-    }
-
-    public  Article findArticleByArticleIdAndMemberIdOrThrow(Long articleId, Long memberId) {
-        return articleRepository.findArticleByIdAndMemberId(articleId, memberId)
-                .orElseThrow(() -> new DomainException(ArticleErrorCode.NOT_FOUND_TO_MEMBER));
-    }
-}
